deleting outdated files and moving movie class
