--- file.csv
@@ -1,6 +1,0 @@
-MOVIE,TICKETS
-movie1,62.0
-movie2,91.0
-movie3,91.0
-movie4,96.0
-movie5,83.0
