--- localVars.py
@@ -1,2 +1,0 @@
-# this file is going to be for any local environmental files you will need an exmaple could be
-image1path = "/Users/seanryan/Documents/cs4125-Project/Images/Image1"
