--- templates/login_TEST_CLASS.html
@@ -1,10 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-
-        <p>THIS IS A LOGIN SUCCESS PAGE. ONLY FOR TESTING. THIS WILL BE DELETED AND A PROPER PAGE MADE</p>
-
-        
-{% endblock %}
-
-
